+++ src/api/api.ts
@@ -8,6 +8,289 @@
  * ---------------------------------------------------------------
  */
 
+export enum CountryCodeEnum {
+  BGD = 'BGD',
+  BEL = 'BEL',
+  BFA = 'BFA',
+  BGR = 'BGR',
+  BIH = 'BIH',
+  BRB = 'BRB',
+  WLF = 'WLF',
+  BLM = 'BLM',
+  BMU = 'BMU',
+  BRN = 'BRN',
+  BOL = 'BOL',
+  BHR = 'BHR',
+  BDI = 'BDI',
+  BEN = 'BEN',
+  BTN = 'BTN',
+  JAM = 'JAM',
+  BVT = 'BVT',
+  BWA = 'BWA',
+  WSM = 'WSM',
+  BES = 'BES',
+  BRA = 'BRA',
+  BHS = 'BHS',
+  JEY = 'JEY',
+  BLR = 'BLR',
+  BLZ = 'BLZ',
+  RUS = 'RUS',
+  RWA = 'RWA',
+  SRB = 'SRB',
+  TLS = 'TLS',
+  REU = 'REU',
+  TKM = 'TKM',
+  TJK = 'TJK',
+  ROU = 'ROU',
+  TKL = 'TKL',
+  GNB = 'GNB',
+  GUM = 'GUM',
+  GTM = 'GTM',
+  SGS = 'SGS',
+  GRC = 'GRC',
+  GNQ = 'GNQ',
+  GLP = 'GLP',
+  JPN = 'JPN',
+  GUY = 'GUY',
+  GGY = 'GGY',
+  GUF = 'GUF',
+  GEO = 'GEO',
+  GRD = 'GRD',
+  GBR = 'GBR',
+  GAB = 'GAB',
+  SLV = 'SLV',
+  GIN = 'GIN',
+  GMB = 'GMB',
+  GRL = 'GRL',
+  GIB = 'GIB',
+  GHA = 'GHA',
+  OMN = 'OMN',
+  TUN = 'TUN',
+  JOR = 'JOR',
+  HRV = 'HRV',
+  HTI = 'HTI',
+  HUN = 'HUN',
+  HKG = 'HKG',
+  HND = 'HND',
+  HMD = 'HMD',
+  VEN = 'VEN',
+  PRI = 'PRI',
+  PSE = 'PSE',
+  PLW = 'PLW',
+  PRT = 'PRT',
+  SJM = 'SJM',
+  PRY = 'PRY',
+  IRQ = 'IRQ',
+  PAN = 'PAN',
+  PYF = 'PYF',
+  PNG = 'PNG',
+  PER = 'PER',
+  PAK = 'PAK',
+  PHL = 'PHL',
+  PCN = 'PCN',
+  POL = 'POL',
+  SPM = 'SPM',
+  ZMB = 'ZMB',
+  ESH = 'ESH',
+  EST = 'EST',
+  EGY = 'EGY',
+  ZAF = 'ZAF',
+  ECU = 'ECU',
+  ITA = 'ITA',
+  VNM = 'VNM',
+  SLB = 'SLB',
+  ETH = 'ETH',
+  SOM = 'SOM',
+  ZWE = 'ZWE',
+  SAU = 'SAU',
+  ESP = 'ESP',
+  ERI = 'ERI',
+  MNE = 'MNE',
+  MDA = 'MDA',
+  MDG = 'MDG',
+  MAF = 'MAF',
+  MAR = 'MAR',
+  MCO = 'MCO',
+  UZB = 'UZB',
+  MMR = 'MMR',
+  MLI = 'MLI',
+  MAC = 'MAC',
+  MNG = 'MNG',
+  MHL = 'MHL',
+  MKD = 'MKD',
+  MUS = 'MUS',
+  MLT = 'MLT',
+  MWI = 'MWI',
+  MDV = 'MDV',
+  MTQ = 'MTQ',
+  MNP = 'MNP',
+  MSR = 'MSR',
+  MRT = 'MRT',
+  IMN = 'IMN',
+  UGA = 'UGA',
+  TZA = 'TZA',
+  MYS = 'MYS',
+  MEX = 'MEX',
+  ISR = 'ISR',
+  FRA = 'FRA',
+  IOT = 'IOT',
+  SHN = 'SHN',
+  FIN = 'FIN',
+  FJI = 'FJI',
+  FLK = 'FLK',
+  FSM = 'FSM',
+  FRO = 'FRO',
+  NIC = 'NIC',
+  NLD = 'NLD',
+  NOR = 'NOR',
+  NAM = 'NAM',
+  VUT = 'VUT',
+  NCL = 'NCL',
+  NER = 'NER',
+  NFK = 'NFK',
+  NGA = 'NGA',
+  NZL = 'NZL',
+  NPL = 'NPL',
+  NRU = 'NRU',
+  NIU = 'NIU',
+  COK = 'COK',
+  XKX = 'XKX',
+  CIV = 'CIV',
+  CHE = 'CHE',
+  COL = 'COL',
+  CHN = 'CHN',
+  CMR = 'CMR',
+  CHL = 'CHL',
+  CCK = 'CCK',
+  CAN = 'CAN',
+  COG = 'COG',
+  CAF = 'CAF',
+  COD = 'COD',
+  CZE = 'CZE',
+  CYP = 'CYP',
+  CXR = 'CXR',
+  CRI = 'CRI',
+  CUW = 'CUW',
+  CPV = 'CPV',
+  CUB = 'CUB',
+  SWZ = 'SWZ',
+  SYR = 'SYR',
+  SXM = 'SXM',
+  KGZ = 'KGZ',
+  KEN = 'KEN',
+  SSD = 'SSD',
+  SUR = 'SUR',
+  KIR = 'KIR',
+  KHM = 'KHM',
+  KNA = 'KNA',
+  COM = 'COM',
+  STP = 'STP',
+  SVK = 'SVK',
+  KOR = 'KOR',
+  SVN = 'SVN',
+  PRK = 'PRK',
+  KWT = 'KWT',
+  SEN = 'SEN',
+  SMR = 'SMR',
+  SLE = 'SLE',
+  SYC = 'SYC',
+  KAZ = 'KAZ',
+  CYM = 'CYM',
+  SGP = 'SGP',
+  SWE = 'SWE',
+  SDN = 'SDN',
+  DOM = 'DOM',
+  DMA = 'DMA',
+  DJI = 'DJI',
+  DNK = 'DNK',
+  VGB = 'VGB',
+  DEU = 'DEU',
+  YEM = 'YEM',
+  DZA = 'DZA',
+  USA = 'USA',
+  URY = 'URY',
+  MYT = 'MYT',
+  UMI = 'UMI',
+  LBN = 'LBN',
+  LCA = 'LCA',
+  LAO = 'LAO',
+  TUV = 'TUV',
+  TWN = 'TWN',
+  TTO = 'TTO',
+  TUR = 'TUR',
+  LKA = 'LKA',
+  LIE = 'LIE',
+  LVA = 'LVA',
+  TON = 'TON',
+  LTU = 'LTU',
+  LUX = 'LUX',
+  LBR = 'LBR',
+  LSO = 'LSO',
+  THA = 'THA',
+  ATF = 'ATF',
+  TGO = 'TGO',
+  TCD = 'TCD',
+  TCA = 'TCA',
+  LBY = 'LBY',
+  VAT = 'VAT',
+  VCT = 'VCT',
+  ARE = 'ARE',
+  AND = 'AND',
+  ATG = 'ATG',
+  AFG = 'AFG',
+  AIA = 'AIA',
+  VIR = 'VIR',
+  ISL = 'ISL',
+  IRN = 'IRN',
+  ARM = 'ARM',
+  ALB = 'ALB',
+  AGO = 'AGO',
+  ATA = 'ATA',
+  ASM = 'ASM',
+  ARG = 'ARG',
+  AUS = 'AUS',
+  AUT = 'AUT',
+  ABW = 'ABW',
+  IND = 'IND',
+  ALA = 'ALA',
+  AZE = 'AZE',
+  IRL = 'IRL',
+  IDN = 'IDN',
+  UKR = 'UKR',
+  QAT = 'QAT',
+  MOZ = 'MOZ',
+}
+
+export interface CreateTenantDto {
+  name: string;
+  document: string;
+
+  /** @example BRA */
+  countryCode: CountryCodeEnum;
+}
+
+export interface TenantEntityDto {
+  id: string;
+  name: string;
+  document: string;
+
+  /** @example BRA */
+  countryCode: CountryCodeEnum;
+  wallets: string[];
+  client?: object;
+  clientId?: object;
+  info: object;
+
+  /** @format date-time */
+  createdAt?: string;
+
+  /** @format date-time */
+  updatedAt?: string;
+
+  /** @format date-time */
+  deletedAt?: string;
+}
+
 export interface HttpExceptionDto {
   timestamp: string;
   path: string;
@@ -15,6 +298,60 @@ export interface HttpExceptionDto {
   statusCode: number;
   message: string;
   data?: object;
+}
+
+export enum OrderByEnum {
+  ASC = 'ASC',
+  DESC = 'DESC',
+}
+
+export interface PaginationMetaDto {
+  /** @example 1 */
+  itemCount: number;
+
+  /** @example 1 */
+  totalItems?: number;
+
+  /** @example 1 */
+  itemsPerPage: number;
+
+  /** @example 1 */
+  totalPages?: number;
+
+  /** @example 1 */
+  currentPage: number;
+}
+
+export interface PaginationLinksDto {
+  /** @example http://example.com?page=1 */
+  first?: string;
+
+  /** @example http://example.com?page=1 */
+  prev?: string;
+
+  /** @example http://example.com?page=2 */
+  next?: string;
+
+  /** @example http://example.com?page=3 */
+  last?: string;
+}
+
+export interface AbstractBase {
+  items: TenantEntityDto[];
+  meta: PaginationMetaDto;
+  links?: PaginationLinksDto;
+}
+
+export interface UpdateTenantProfileDto {
+  name?: string;
+}
+
+export interface UpdateTenantDto {
+  name?: string;
+  document?: string;
+
+  /** @example BRA */
+  countryCode?: CountryCodeEnum;
 }
 
 export interface RequestConfirmationEmailDto {
@@ -42,6 +379,12 @@ export interface ResetPasswordDto {
   token: string;
 }
 
+export interface SignInResponseDto {
+  token: string;
+  refreshToken: string;
+  data: object;
+}
+
 export interface LoginUserDto {
   /** @example user@example.com */
   email: string;
@@ -53,12 +396,6 @@ export interface LoginUserDto {
   tenantId?: string;
 }
 
-export interface SignInResponseDto {
-  token: string;
-  refreshToken: string;
-  data: object;
-}
-
 export interface RefreshTokenDto {
   /** @example eyJhbGciOiJSUzI1NiIsInR5cCI6InJlZnJlc2gifQ.eyJzdWIiOiI4NGViM2Q2NC01OTk4LTRjM2UtODliMS0yZTgyZmQ2NDRjMDciLCJpc3MiOiIwMDAwMDAwMC0wMDAwLTAwMDAtMDAwMC0wMDAwMDAwMDAwMDEiLCJhdWQiOiIwMDAwMDAwMC0wMDAwLTAwMDAtMDAwMC0wMDAwMDAwMDAwMDEiLCJjb21wYW55SWQiOiIwMDAwMDAwMC0wMDAwLTAwMDAtMDAwMC0wMDAwMDAwMDAwMDEiLCJ0b2tlbkhhc2giOiI5NGFkMzFjMjcwMWI0NzQxNjk5ZjI1YTA4NWM0NmQyNDE4ZTk2MjNhYTcyYjM3MGZhZWZkNWJhYWYzYzM3YzgzIiwiaWF0IjoxNjU3MzEwMzQwLCJleHAiOjE2NTc0ODMxNDB9.epg4Grc6LkJMMxNlFbsdEc5Rq2nupXZTsFQeD-JgKZMWg4ey_ialD-mmSBehHD_xFF4Ho_BbWG1ld9aCuQBQcft4zxDfMdxogM5fMfdhAnAccM7a4J0NyKECzYHhIDe_jnrpT4QVHs4x78NIookQiKguANuSa7dNViFphpC9K5LRcxVwS6sEqT8 */
   refreshToken: string;
@@ -67,6 +404,206 @@ export interface RefreshTokenDto {
 export interface RefreshTokenResponseDto {
   token: string;
   refreshToken: string;
+}
+
+export namespace Tenant {
+  /**
+   * No description
+   * @tags Tenant
+   * @name Create
+   * @request POST:/tenant
+   * @secure
+   */
+  export namespace Create {
+    export type RequestParams = {};
+    export type RequestQuery = {};
+    export type RequestBody = CreateTenantDto;
+    export type RequestHeaders = {};
+    export type ResponseBody = TenantEntityDto;
+  }
+  /**
+   * No description
+   * @tags Tenant
+   * @name FindAll
+   * @request GET:/tenant
+   * @secure
+   */
+  export namespace FindAll {
+    export type RequestParams = {};
+    export type RequestQuery = {
+      page?: number;
+      limit?: number;
+      search?: string;
+      sortBy?: string;
+      orderBy?: OrderByEnum;
+    };
+    export type RequestBody = never;
+    export type RequestHeaders = {};
+    export type ResponseBody = AbstractBase;
+  }
+  /**
+   * No description
+   * @tags Tenant
+   * @name FindOne
+   * @request GET:/tenant/{id}
+   * @secure
+   */
+  export namespace FindOne {
+    export type RequestParams = { id: string };
+    export type RequestQuery = {};
+    export type RequestBody = never;
+    export type RequestHeaders = {};
+    export type ResponseBody = any;
+  }
+  /**
+   * No description
+   * @tags Tenant
+   * @name Update
+   * @request PUT:/tenant/{id}
+   * @secure
+   */
+  export namespace Update {
+    export type RequestParams = { id: string };
+    export type RequestQuery = {};
+    export type RequestBody = UpdateTenantDto;
+    export type RequestHeaders = {};
+    export type ResponseBody = any;
+  }
+  /**
+   * No description
+   * @tags Tenant
+   * @name Remove
+   * @request DELETE:/tenant/{id}
+   * @secure
+   */
+  export namespace Remove {
+    export type RequestParams = { id: string };
+    export type RequestQuery = {};
+    export type RequestBody = never;
+    export type RequestHeaders = {};
+    export type ResponseBody = any;
+  }
+  /**
+   * No description
+   * @tags Tenant
+   * @name UpdateProfile
+   * @request PUT:/tenant/profile/{id}
+   * @secure
+   */
+  export namespace UpdateProfile {
+    export type RequestParams = { id: string };
+    export type RequestQuery = {};
+    export type RequestBody = UpdateTenantProfileDto;
+    export type RequestHeaders = {};
+    export type ResponseBody = any;
+  }
+}
+
+export namespace Auth {
+  /**
+   * No description
+   * @tags Authentication
+   * @name VerifySignUp
+   * @request GET:/auth/verify-sign-up
+   */
+  export namespace VerifySignUp {
+    export type RequestParams = {};
+    export type RequestQuery = { email: string; token: string };
+    export type RequestBody = never;
+    export type RequestHeaders = {};
+    export type ResponseBody = any;
+  }
+  /**
+   * No description
+   * @tags Authentication
+   * @name RequestConfirmationEmail
+   * @request POST:/auth/request-confirmation-email
+   */
+  export namespace RequestConfirmationEmail {
+    export type RequestParams = {};
+    export type RequestQuery = {};
+    export type RequestBody = RequestConfirmationEmailDto;
+    export type RequestHeaders = {};
+    export type ResponseBody = void;
+  }
+  /**
+   * No description
+   * @tags Authentication
+   * @name RequestPasswordReset
+   * @request POST:/auth/request-password-reset
+   */
+  export namespace RequestPasswordReset {
+    export type RequestParams = {};
+    export type RequestQuery = {};
+    export type RequestBody = RequestPasswordResetDto;
+    export type RequestHeaders = {};
+    export type ResponseBody = void;
+  }
+  /**
+   * No description
+   * @tags Authentication
+   * @name ResetPassword
+   * @request POST:/auth/reset-password
+   */
+  export namespace ResetPassword {
+    export type RequestParams = {};
+    export type RequestQuery = {};
+    export type RequestBody = ResetPasswordDto;
+    export type RequestHeaders = {};
+    export type ResponseBody = SignInResponseDto;
+  }
+  /**
+   * No description
+   * @tags Authentication
+   * @name SignIn
+   * @request POST:/auth/signin
+   */
+  export namespace SignIn {
+    export type RequestParams = {};
+    export type RequestQuery = {};
+    export type RequestBody = LoginUserDto;
+    export type RequestHeaders = {};
+    export type ResponseBody = SignInResponseDto;
+  }
+  /**
+   * No description
+   * @tags Authentication
+   * @name RefreshToken
+   * @request POST:/auth/refresh-token
+   */
+  export namespace RefreshToken {
+    export type RequestParams = {};
+    export type RequestQuery = {};
+    export type RequestBody = RefreshTokenDto;
+    export type RequestHeaders = {};
+    export type ResponseBody = RefreshTokenResponseDto;
+  }
+  /**
+   * No description
+   * @tags Authentication
+   * @name LogOut
+   * @request POST:/auth/logout
+   */
+  export namespace LogOut {
+    export type RequestParams = {};
+    export type RequestQuery = {};
+    export type RequestBody = never;
+    export type RequestHeaders = {};
+    export type ResponseBody = void;
+  }
+  /**
+   * No description
+   * @tags Authentication
+   * @name GetJwks
+   * @request GET:/auth/jwks.json
+   */
+  export namespace GetJwks {
+    export type RequestParams = {};
+    export type RequestQuery = {};
+    export type RequestBody = never;
+    export type RequestHeaders = {};
+    export type ResponseBody = void;
+  }
 }
 
 import axios, { AxiosInstance, AxiosRequestConfig, AxiosResponse, ResponseType } from 'axios';
@@ -196,6 +733,115 @@ export class HttpClient<SecurityDataType = unknown> {
  * @contact
  */
 export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDataType> {
+  tenant = {
+    /**
+     * No description
+     *
+     * @tags Tenant
+     * @name Create
+     * @request POST:/tenant
+     * @secure
+     */
+    create: (data: CreateTenantDto, params: RequestParams = {}) =>
+      this.request<TenantEntityDto, HttpExceptionDto>({
+        path: `/tenant`,
+        method: 'POST',
+        body: data,
+        secure: true,
+        type: ContentType.Json,
+        format: 'json',
+        ...params,
+      }),
+
+    /**
+     * No description
+     *
+     * @tags Tenant
+     * @name FindAll
+     * @request GET:/tenant
+     * @secure
+     */
+    findAll: (
+      query?: { page?: number; limit?: number; search?: string; sortBy?: string; orderBy?: OrderByEnum },
+      params: RequestParams = {},
+    ) =>
+      this.request<AbstractBase, HttpExceptionDto>({
+        path: `/tenant`,
+        method: 'GET',
+        query: query,
+        secure: true,
+        format: 'json',
+        ...params,
+      }),
+
+    /**
+     * No description
+     *
+     * @tags Tenant
+     * @name FindOne
+     * @request GET:/tenant/{id}
+     * @secure
+     */
+    findOne: (id: string, params: RequestParams = {}) =>
+      this.request<any, HttpExceptionDto>({
+        path: `/tenant/${id}`,
+        method: 'GET',
+        secure: true,
+        ...params,
+      }),
+
+    /**
+     * No description
+     *
+     * @tags Tenant
+     * @name Update
+     * @request PUT:/tenant/{id}
+     * @secure
+     */
+    update: (id: string, data: UpdateTenantDto, params: RequestParams = {}) =>
+      this.request<any, HttpExceptionDto>({
+        path: `/tenant/${id}`,
+        method: 'PUT',
+        body: data,
+        secure: true,
+        type: ContentType.Json,
+        ...params,
+      }),
+
+    /**
+     * No description
+     *
+     * @tags Tenant
+     * @name Remove
+     * @request DELETE:/tenant/{id}
+     * @secure
+     */
+    remove: (id: string, params: RequestParams = {}) =>
+      this.request<any, HttpExceptionDto>({
+        path: `/tenant/${id}`,
+        method: 'DELETE',
+        secure: true,
+        ...params,
+      }),
+
+    /**
+     * No description
+     *
+     * @tags Tenant
+     * @name UpdateProfile
+     * @request PUT:/tenant/profile/{id}
+     * @secure
+     */
+    updateProfile: (id: string, data: UpdateTenantProfileDto, params: RequestParams = {}) =>
+      this.request<any, HttpExceptionDto>({
+        path: `/tenant/profile/${id}`,
+        method: 'PUT',
+        body: data,
+        secure: true,
+        type: ContentType.Json,
+        ...params,
+      }),
+  };
   auth = {
     /**
      * No description
@@ -252,11 +898,12 @@ export class Api<SecurityDataType extends unknown> extends HttpClient<SecurityDa
      * @request POST:/auth/reset-password
      */
     resetPassword: (data: ResetPasswordDto, params: RequestParams = {}) =>
-      this.request<void, HttpExceptionDto>({
+      this.request<SignInResponseDto, HttpExceptionDto>({
         path: `/auth/reset-password`,
         method: 'POST',
         body: data,
         type: ContentType.Json,
+        format: 'json',
         ...params,
       }),