+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.3](https://github.com/TropixInc/pixwayid-sdk/compare/v0.0.1...v0.0.3) (2022-07-21)
+
+
+### Features
+
+* add ts disable to api generated file ([01ff69c](https://github.com/TropixInc/pixwayid-sdk/commit/01ff69c7a5597c90b99d920a20c3a30f406b8ecf))
+
+
+### Bug Fixes
+
+* set auth header to generated api client ([67e0af2](https://github.com/TropixInc/pixwayid-sdk/commit/67e0af2a994ce2a1ac7be093e60eb7fd7aabbb74))
+
+### [0.0.2](https://github.com/TropixInc/pixwayid-sdk/compare/v0.0.1...v0.0.2) (2022-07-21)
+
+
+### Features
+
+* add ts disable to api generated file ([01ff69c](https://github.com/TropixInc/pixwayid-sdk/commit/01ff69c7a5597c90b99d920a20c3a30f406b8ecf))
+
+
+### Bug Fixes
+
+* set auth header to generated api client ([67e0af2](https://github.com/TropixInc/pixwayid-sdk/commit/67e0af2a994ce2a1ac7be093e60eb7fd7aabbb74))
+
 ### 0.0.1 (2022-07-18)